requirements.txt
python-dotenv==1.0.0
clickhouse-connect==0.7.19
requests==2.31.0
aiohttp==3.9.1
pymongo==4.14.0
dnspython==2.7.0
flask==3.0.0
